requirements.txt
# Dependencias para fine-tuning NLLB Awajún-Español

# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0

# MLflow tracking
mlflow>=2.5.0
PyYAML>=6.0

# Evaluation metrics
sacrebleu>=2.3.0
sacremoses>=0.0.53

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.65.0

# Optional: for better performance
# accelerate>=0.20.0  # Para multi-GPU
# datasets[audio]>=2.10.0  # Si necesitas audio

# Development
jupyter>=1.0.0
ipython>=8.0.0